Increase cache versioning precision to allow subsecond difference.

+++ lib/moebooru/cache_helper.rb
@@ -1,13 +1,13 @@
 module Moebooru
   module CacheHelper
     def expire(_options = {})
-      Rails.cache.write("$cache_version", Time.now.to_i)
+      Rails.cache.write("$cache_version", Time.now.to_f)
     end
 
     def expire_tag_version
       # $tag_version is bumped when the type of a tag is changed in Tags, if
       # a new tag is created, or if a tag's post_count becomes nonzero.
-      Rails.cache.write("$tag_version", Time.now.to_i)
+      Rails.cache.write("$tag_version", Time.now.to_f)
     end
 
     module_function :expire, :expire_tag_version
